refactor: share local asset path validation
